requirements.txt
androguard==3.3.5
asn1crypto==1.5.1
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.8.0
fonttools==4.38.0
future==0.18.2
huggingface-hub==0.11.0
idna==3.4
ipython==8.6.0
jedi==0.18.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
networkx==2.8.8
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
stack-data==0.6.1
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5